feat(admin verification): verify admin before executing delete and update on entities

+++ src/main/java/com/vdoshi3/controller/CommentControllerImp.java
@@ -2,8 +2,11 @@ package com.vdoshi3.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,8 +17,11 @@ import org.springframework.web.bind.annotation.RestController;
 import com.fasterxml.jackson.annotation.JsonView;
 import com.vdoshi3.entity.Comment;
 import com.vdoshi3.entity.View;
+import com.vdoshi3.exception.InvalidCredentialsException;
 import com.vdoshi3.exception.ResourceNotFoundException;
 import com.vdoshi3.service.CommentService;
+import com.vdoshi3.utils.DecodedToken;
+
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
@@ -39,7 +45,7 @@ public class CommentControllerImp implements CommentController {
 			@RequestBody Comment comment) throws ResourceNotFoundException {
 		return service.create(mid, uid, comment);
 	}
-	
+
 	@JsonView(View.Public.class)
 	@Override
 	@RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
@@ -75,8 +81,17 @@ public class CommentControllerImp implements CommentController {
 	@ApiResponses(value = { @ApiResponse(code = 200, message = "Success"),
 			@ApiResponse(code = 404, message = "Not Found"),
 			@ApiResponse(code = 500, message = "Internal Server Error") })
-	public void delete(@PathVariable("cid") String cid) throws ResourceNotFoundException {
-		service.delete(cid);
+	public void delete(@PathVariable("cid") String cid, @ModelAttribute("requestor") DecodedToken requestor)
+			throws ResourceNotFoundException, InvalidCredentialsException {
+		if (requestor.getSubject().equals("admin")) {
+			service.delete(cid);
+		} else {
+			throw new InvalidCredentialsException();
+		}
 	}
 
+	@ModelAttribute("requestor")
+	public DecodedToken getDecodedToken(HttpServletRequest request) {
+		return (DecodedToken) request.getAttribute("requestor");
+	}
 }
