+++ src/main/java/com/vdoshi3/controller/RatingControllerImp.java
@@ -2,8 +2,11 @@ package com.vdoshi3.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,8 +17,11 @@ import org.springframework.web.bind.annotation.RestController;
 import com.fasterxml.jackson.annotation.JsonView;
 import com.vdoshi3.entity.Rating;
 import com.vdoshi3.entity.View;
+import com.vdoshi3.exception.InvalidCredentialsException;
 import com.vdoshi3.exception.ResourceNotFoundException;
 import com.vdoshi3.service.RatingService;
+import com.vdoshi3.utils.DecodedToken;
+
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
@@ -37,7 +43,7 @@ public class RatingControllerImp implements RatingController {
 			@ApiResponse(code = 500, message = "Internal Server Error") })
 	public Rating create(@RequestParam("mid") String mid, @RequestParam("uid") String uid, @RequestBody Rating rating)
 			throws ResourceNotFoundException {
-			return service.create(mid, uid, rating);
+		return service.create(mid, uid, rating);
 	}
 
 	@JsonView(View.Public.class)
@@ -76,8 +82,17 @@ public class RatingControllerImp implements RatingController {
 	@ApiResponses(value = { @ApiResponse(code = 200, message = "Success"),
 			@ApiResponse(code = 404, message = "Not Found"),
 			@ApiResponse(code = 500, message = "Internal Server Error") })
-	public void delete(@PathVariable("rid") String rid) throws ResourceNotFoundException {
-		service.delete(rid);
+	public void delete(@PathVariable("rid") String rid, @ModelAttribute("requestor") DecodedToken requestor)
+			throws ResourceNotFoundException, InvalidCredentialsException {
+		if (requestor.getSubject().equals("admin")) {
+			service.delete(rid);
+		} else {
+			throw new InvalidCredentialsException();
+		}
 	}
 
+	@ModelAttribute("requestor")
+	public DecodedToken getDecodedToken(HttpServletRequest request) {
+		return (DecodedToken) request.getAttribute("requestor");
+	}
 }
